requirements.txt
absl-py==2.1.0
aiobotocore==2.13.2
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aioitertools==0.11.0
aiosignal==1.3.1
anndata==0.10.8
antlr4-python3-runtime==4.9.3
array_api_compat==1.8
asciitree==0.3.3
BiocFrame==0.5.11
biocutils==0.1.6
botocore==1.34.131
captum==0.7.0
circlify==0.15.0
click==8.1.7
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
docker-pycreds==0.4.0
docstring_parser==0.16
fasteners==0.19
fonttools==4.53.1
frozenlist==1.4.1
GenomicRanges==0.4.34
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.65.4
h5py==3.12.1
hnswlib==0.8.0
hpack==4.0.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
igraph==0.11.6
IRanges==0.2.12
jmespath==1.0.1
joblib==1.4.2
jsonargparse==4.32.0
kiwisolver==1.4.5
legacy-api-wrap==1.4
leidenalg==0.10.2
lightning 
lightning-utilities 
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
matplotlib==3.9.1.post1
mdurl==0.1.2
mopsy==0.2.8
multidict==6.0.5
natsort==8.4.0
ncls==0.0.68
numba==0.60.0
numcodecs==0.13.0
numpy
obonet==1.1.0
omegaconf==2.3.0
pandas==2.2.2
patsy==0.5.6
polars==1.12.0
POT==0.9.4
protobuf==4.25.4
pyarrow==17.0.0
pynndescent==0.5.13
pyparsing==3.1.2
python-dotenv==1.0.1
pytorch-lightning==2.4.0
rich==13.7.1
rootutils==1.0.7
s3fs==2024.6.1
scanpy==1.10.2
scikit-learn==1.5.1
scimilarity
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==2.12.0
session_info==1.0.0
setproctitle==1.3.3
smmap==5.0.1
statsmodels==0.14.2
stdlib-list==0.10.0
SummarizedExperiment==0.4.6
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
texttable==1.7.0
threadpoolctl==3.5.0
tiledb==0.31.1
torch==2.4.0
torchmetrics
torchvision==0.19.0
tqdm 
triton==3.0.0
typeshed_client==2.7.0
tzdata==2024.1
umap-learn==0.5.6
wandb==0.17.6
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zarr==2.18.2
zstandard==0.23.0
imblearn
-e .